+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Allow event custom range filters in the past [#4450](https://github.com/rokwire/illinois-app/issues/4450).
 - Added getTextStyleEx to UiTextStyles.
+- Added scheme parameter to UrlUtils.fixUrl.
 
 ## [1.7.4] - 2024-11-07
 ### Fixed